Bug Fix: tickets can now be marked complete or incomplete
Fixed an issue in which sending PUT or DELETE requests to /complete was
not updating the completed status.

As a side note, as this is essentially a toggle, couldn't this simply be
one endpoint instead of two with different methods? I would think that
the DELETE method would be reserved for deleting a ticket.